Added Integration Tests For The Admin Controller In The Gateway API
I added all the necessary integration tests for the admin controller that was added in the previous commit to the gateway api. Those tests check the CRUD procedures that the admin can access through the gateway app(it communicates with the Auth Library Admin Controller for that purpose). When it comes to the code I just changed some return status codes that the admin controller of the auth microservice to notfound instead of bad request. The admincontroller of the gateway api also changed in some parts after issues were found through testing.

+++ EshopApp.TestUtilitiesLibrary/EmailUtilities.cs
@@ -1,30 +1,34 @@
 ﻿namespace EshopApp.TestUtilitiesLibrary;
 public class EmailUtilities
 {
-    public static List<string>? ReadLastEmailFile()
+    public static string? ReadLastEmailFile(bool deleteEmailFile)
     {
         string directoryPath = @"C:\ProgramData\Changemaker Studios\Papercut SMTP\Incoming";
 
         List<string> emlFiles = Directory.GetFiles(directoryPath, "*.eml")
-                                .OrderByDescending(f => new FileInfo(f).LastWriteTime)
-                                .ToList();
+            .OrderByDescending(f => new FileInfo(f).LastWriteTime)
+            .ToList();
 
         string? lastEmailFile = emlFiles.FirstOrDefault();
         if (lastEmailFile == null)
             return null;
 
-        List<string> emailFileLines = File.ReadAllLines(lastEmailFile).ToList();
-        return emailFileLines;
+        string emailFileContent = File.ReadAllText(lastEmailFile);
+
+        if (deleteEmailFile)
+            File.Delete(lastEmailFile);
+
+        return emailFileContent;
     }
 
-    public static string? GetLastEmailLink()
+    public static string? GetLastEmailLink(bool deleteEmailFile)
     {
         string directoryPath = @"C:\ProgramData\Changemaker Studios\Papercut SMTP\Incoming";
 
         // Get all .eml files in the directory
         List<string> emlFiles = Directory.GetFiles(directoryPath, "*.eml")
-                                .OrderByDescending(f => new FileInfo(f).LastWriteTime)
-                                .ToList();
+            .OrderByDescending(f => new FileInfo(f).LastWriteTime)
+            .ToList();
 
         // Get the last one
         string? lastEmailFile = emlFiles.FirstOrDefault();
@@ -46,11 +50,11 @@ public class EmailUtilities
         link = link.Replace("=3D", "=");
 
         // Delete the file now that we have the link
-        File.Delete(lastEmailFile);
+        if (deleteEmailFile)
+            File.Delete(lastEmailFile);
 
         return link;
     }
-
 
     public static void DeleteAllEmailFiles()
     {
